+++ examples/bulk.rs
@@ -1,10 +1,8 @@
 use indicatif::ProgressBar;
 use once_cell::sync::Lazy;
-use std::borrow::Cow;
 use std::env;
 use tiberius::numeric::Numeric;
-use tiberius::time::{Date, DateTime, DateTime2, DateTimeOffset, SmallDateTime, Time};
-use tiberius::xml::XmlData;
+use tiberius::time::{Date, DateTime, DateTime2, DateTimeOffset, Time};
 use tiberius::{Client, ColumnData, Config, IntoSql, TokenRow};
 use tokio::net::TcpStream;
 use tokio_util::compat::TokioAsyncWriteCompatExt;
