+++ src/tds/codec/column_data.rs
@@ -1,6 +1,5 @@
 mod binary;
 mod bit;
-mod buf;
 #[cfg(feature = "tds73")]
 mod date;
 #[cfg(feature = "tds73")]
@@ -29,7 +28,6 @@ use crate::{
     tds::{codec::TypeInfoInner, time::DateTime, time::SmallDateTime, xml::XmlData, Numeric},
     SqlReadBytes,
 };
-pub(crate) use buf::BufColumnData;
 use bytes::{BufMut, BytesMut};
 use encoding::EncoderTrap;
 use std::borrow::{BorrowMut, Cow};
@@ -404,241 +402,6 @@ impl<'a> ColumnData<'a> {
             (v, ref ti) => Err(crate::Error::BulkInput(
                 format!("invalid data type, expecting {:?} but found {:?}", ti, v).into(),
             ))?,
-        }
-
-        Ok(())
-    }
-}
-
-impl<'a> Encode<BufColumnData<'a>> for ColumnData<'a> {
-    fn encode(self, dst: &mut BufColumnData<'a>) -> crate::Result<()> {
-        match self {
-            ColumnData::Bit(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Bitn as u8, 1, 1][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_u8(val as u8);
-            }
-            ColumnData::U8(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Intn as u8, 1, 1][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_u8(val);
-            }
-            ColumnData::I16(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Intn as u8, 2, 2][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_i16_le(val);
-            }
-            ColumnData::I32(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Intn as u8, 4, 4][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_i32_le(val);
-            }
-            ColumnData::I64(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Intn as u8, 8, 8][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_i64_le(val);
-            }
-            ColumnData::F32(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Floatn as u8, 4, 4][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_f32_le(val);
-            }
-            ColumnData::F64(Some(val)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Floatn as u8, 8, 8][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                dst.put_f64_le(val);
-            }
-            ColumnData::Guid(Some(uuid)) => {
-                if dst.write_headers {
-                    let header = [&[VarLenType::Guid as u8, 16, 16][..]].concat();
-                    dst.extend_from_slice(&header);
-                }
-
-                let mut data = *uuid.as_bytes();
-                super::guid::reorder_bytes(&mut data);
-                dst.extend_from_slice(&data);
-            }
-            ColumnData::String(Some(ref s)) if s.len() <= 4000 => {
-                if dst.write_headers {
-                    dst.put_u8(VarLenType::NVarchar as u8);
-                    dst.put_u16_le(8000);
-                    dst.extend_from_slice(&[0u8; 5][..]);
-                }
-
-                let mut length = 0u16;
-                let len_pos = dst.len();
-
-                dst.put_u16_le(length);
-
-                for chr in s.encode_utf16() {
-                    length += 1;
-                    dst.put_u16_le(chr);
-                }
-
-                let dst: &mut [u8] = dst.borrow_mut();
-                let bytes = (length * 2).to_le_bytes(); // u16, two bytes
-
-                for (i, byte) in bytes.iter().enumerate() {
-                    dst[len_pos + i] = *byte;
-                }
-            }
-            ColumnData::String(Some(ref s)) => {
-                if dst.write_headers {
-                    // length: 0xffff and raw collation
-                    dst.put_u8(VarLenType::NVarchar as u8);
-                    dst.extend_from_slice(&[0xff_u8; 2][..]);
-                    dst.extend_from_slice(&[0u8; 5][..]);
-
-                    // we cannot cheaply predetermine the length of the UCS2 string beforehand
-                    // (2 * bytes(UTF8) is not always right) - so just let the SQL server handle it
-                    dst.put_u64_le(0xfffffffffffffffe_u64);
-                }
-
-                // Write the varchar length
-                let mut length = 0u32;
-                let len_pos = dst.len();
-
-                dst.put_u32_le(length);
-
-                for chr in s.encode_utf16() {
-                    length += 1;
-                    dst.put_u16_le(chr);
-                }
-
-                // PLP_TERMINATOR
-                dst.put_u32_le(0);
-
-                let dst: &mut [u8] = dst.borrow_mut();
-                let bytes = (length * 2).to_le_bytes(); // u32, four bytes
-
-                for (i, byte) in bytes.iter().enumerate() {
-                    dst[len_pos + i] = *byte;
-                }
-            }
-            ColumnData::Binary(Some(bytes)) if bytes.len() <= 8000 => {
-                if dst.write_headers {
-                    dst.put_u8(VarLenType::BigVarBin as u8);
-                    dst.put_u16_le(8000);
-                }
-
-                dst.put_u16_le(bytes.len() as u16);
-                dst.extend(bytes.into_owned());
-            }
-            ColumnData::Binary(Some(bytes)) => {
-                if dst.write_headers {
-                    dst.put_u8(VarLenType::BigVarBin as u8);
-                    // Max length
-                    dst.put_u16_le(0xffff_u16);
-                    // Also the length is unknown
-                    dst.put_u64_le(0xfffffffffffffffe_u64);
-                }
-
-                // We'll write in one chunk, length is the whole bytes length
-                dst.put_u32_le(bytes.len() as u32);
-                // Payload
-                dst.extend(bytes.into_owned());
-                // PLP_TERMINATOR
-                dst.put_u32_le(0);
-            }
-            ColumnData::DateTime(Some(dt)) => {
-                if dst.write_headers {
-                    dst.extend_from_slice(&[VarLenType::Datetimen as u8, 8, 8]);
-                }
-
-                dt.encode(&mut *dst)?;
-            }
-            ColumnData::SmallDateTime(Some(dt)) => {
-                if dst.write_headers {
-                    dst.extend_from_slice(&[VarLenType::Datetimen as u8, 4, 4]);
-                }
-
-                dt.encode(&mut *dst)?;
-            }
-            #[cfg(feature = "tds73")]
-            ColumnData::Time(Some(time)) => {
-                if dst.write_headers {
-                    dst.extend_from_slice(&[VarLenType::Timen as u8, time.scale(), time.len()?]);
-                }
-
-                time.encode(&mut *dst)?;
-            }
-            #[cfg(feature = "tds73")]
-            ColumnData::Date(Some(date)) => {
-                if dst.write_headers {
-                    dst.extend_from_slice(&[VarLenType::Daten as u8, 3]);
-                }
-
-                date.encode(&mut *dst)?;
-            }
-            #[cfg(feature = "tds73")]
-            ColumnData::DateTime2(Some(dt)) => {
-                if dst.write_headers {
-                    let len = dt.time().len()? + 3;
-                    dst.extend_from_slice(&[VarLenType::Datetime2 as u8, dt.time().scale(), len]);
-                }
-
-                dt.encode(&mut *dst)?;
-            }
-            #[cfg(feature = "tds73")]
-            ColumnData::DateTimeOffset(Some(dto)) => {
-                if dst.write_headers {
-                    let headers = &[
-                        VarLenType::DatetimeOffsetn as u8,
-                        dto.datetime2().time().scale(),
-                        dto.datetime2().time().len()? + 5,
-                    ];
-
-                    dst.extend_from_slice(headers);
-                }
-
-                dto.encode(&mut *dst)?;
-            }
-            ColumnData::Xml(Some(xml)) => {
-                if dst.write_headers {
-                    dst.put_u8(VarLenType::Xml as u8);
-                    dst.put_u8(0);
-                }
-                xml.into_owned().encode(&mut *dst)?;
-            }
-            ColumnData::Numeric(Some(num)) => {
-                if dst.write_headers {
-                    let headers = &[
-                        VarLenType::Numericn as u8,
-                        num.len(),
-                        num.precision(),
-                        num.scale(),
-                    ];
-
-                    dst.extend_from_slice(headers);
-                }
-
-                num.encode(&mut *dst)?;
-            }
-            _ => {
-                // None/null
-                dst.put_u8(FixedLenType::Null as u8);
-            }
         }
 
         Ok(())
